Set multitenancy_enabled to False for all jobs
By disabling this option, we limit the ammount of share servers that
will be created when running Tempest tests. We need this as at the
moment we're running out of resources on the compute nodes.

Also, max_fixed_ips_per_port is now properly set, as it was missing
the group label.