Use predefined tenants when running tests
At the moment, we have big issues due to the fact that each time a
share network is needed by a tempest test, a new tenant will be
created, as a result, a new neutron network as well.

For this reason, we quickly get out of VLAN tags.

This change will let devstack configure a predefined list of keystone
accounts, based on the maximum number of parallel tests, thus greatly
reducing the ammount of networks created.