Ensure private IPs are accessible on the devstack VM
At the moment, manila is using the private ips when accessing share
servers.

Devstack is no longer adding a route for the private networks, so
those are no longer accessible.

For the moment, we'll just go back to the old behavior, adding a route
via the public router ip.

As an alternative, we could use floating ips for the share instances
in the 'user share servers' scenario, but that would require restarting
the m-shr service.